pyproject.toml
[project]
name = "universal-icon-click-captcha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.1",
    "pillow>=11.3.0",
]
